chore: rename Package, again
BasePackage => Package
Package => RootPackage

Note: RootPackage could be thought of as "module", but we are reserving
that word due to "go mod"
