+++ packages/jsii-pacmak/lib/targets/golang/types/go-type.ts
@@ -1,13 +1,13 @@
 import { CodeMaker } from 'codemaker';
 import { Type } from 'jsii-reflect';
-import { BasePackage } from '../base-package';
+import { Package } from '../package';
 
 export interface GoEmitter {
   emit(code: CodeMaker): void;
 }
 
 export class GoType {
-  public constructor(public parent: BasePackage, public type: Type) {}
+  public constructor(public parent: Package, public type: Type) {}
 
   public get name() {
     return this.type.name;
